Move location to networkmonitor package.
