+++ core/data-test/src/main/kotlin/com/google/samples/apps/nowinandroid/core/data/test/networkmonitor/AlwaysOnlineNetworkMonitor.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Android Open Source Project
+ * Copyright 2024 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.samples.apps.nowinandroid.core.data.test
+package com.google.samples.apps.nowinandroid.core.data.test.networkmonitor
 
 import com.google.samples.apps.nowinandroid.core.data.util.NetworkMonitor
 import kotlinx.coroutines.flow.Flow
